Trying to fix test issues
